+++ R/data.R
@@ -115,9 +115,13 @@
 #' Brewer palettes
 #'
 #' A list of eleven Brewer palettes containing one to eleven colours that
-#' are readily distinguished by colourblind viewers
+#' are readily distinguished by colourblind viewers, followed by a twelfth
+#' 12-colour palette adapted for colour blindness.
 #'
-#' @source [ColourBrewer2.org](http://colorbrewer2.org/#type=diverging&scheme=RdYlBu&n=3)
+#' @source {
+#'   [ColourBrewer2.org](http://colorbrewer2.org/#type=diverging&scheme=RdYlBu&n=3)
+#'   [Martin Krzywinski](http://mkweb.bcgsc.ca/colorblind/)
+#' }
 #'
 #' @keywords datasets
 "brewer"