+++ NEWS.md
@@ -1,6 +1,7 @@
 # TreeSearch 0.1.2.9000
 
 ## Enhancements
+ - Add new 12-colour colourblind-friendly palette
  - FitchSteps now supports datasets with tips not found in tree
 
 # TreeSearch 0.1.2