+++ NEWS.md
@@ -6,6 +6,7 @@
 ## Enhancements
  - Add new 12-colour colourblind-friendly palette
  - FitchSteps now supports datasets with tips not found in tree
+ - Improve portability of function `ReadTntTree`
 
 # TreeSearch 0.1.2