+++ NEWS.md
@@ -1,7 +1,12 @@
 # TreeSearch 0.1.2.9000
 
+## New features
+ - `RootTree` function allows rooting of tree on incompletely specified outgroup
+ 
 ## Enhancements
-- Improve portability of function `ReadTntTree`
+ - Add new 12-colour colourblind-friendly palette
+ - FitchSteps now supports datasets with tips not found in tree
+ - Improve portability of function `ReadTntTree`
 
 # TreeSearch 0.1.2